+++ writer.go
@@ -159,4 +159,7 @@ type Writer interface {
 
 	// SetColor sets text and background colors. and specify whether text is bold.
 	SetColor(fg, bg Color, bold bool)
+
+	// Sets the colors and display attributes.
+	SetDisplayAttributes(fg, bg Color, attrs ...DisplayAttribute)
 }